Basket products + refactor creating accound data + db_changelog

+++ src/main/java/com/falynsky/smartmarkt/models/BasketProduct.java
@@ -18,7 +18,7 @@ public class BasketProduct {
     int id;
 
     @Column(name = "quantity")
-    float quantity;
+    int quantity;
 
     @Column(name = "quantity_type")
     String quantityType;
